backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
openai==1.3.0
python-dotenv==1.0.0
redis==5.0.1
psycopg2-binary==2.9.9
websockets==12.0
python-socketio==5.10.0
aiofiles==0.7.0
pillow==9.0.0
pydub==0.25.1
langchain==0.0.350
chromadb==0.4.18
tiktoken==0.5.2
PyPDF2==3.0.1
python-docx==0.8.11
pytesseract==0.3.8
SpeechRecognition==3.8.1
numpy==1.24.3 